+++ HISTORY.rst
@@ -2,6 +2,9 @@
 History
 =======
 
+* Move initial ``max_migration.txt`` file creation into a separate management
+  command, ``create-max-migration-files``.
+
 1.0.0 (2020-12-10)
 ------------------